% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{add_dimensions}
\alias{add_dimensions}
\title{Title}
\usage{
add_dimensions(slr, dimensions, output = ".")
}
\arguments{
\item{slr}{SLR data}

\item{dimensions}{New dimensions(columns to add to the Excel file)}

\item{output}{Folder where to write the Excel file with the articles (default is the current directory - '.')}
}
\description{
Title
}
