% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplot_per_year}
\alias{boxplot_per_year}
\title{Title}
\usage{
boxplot_per_year(slr, year.above = 2010, output = ".", save.pdf = TRUE)
}
\arguments{
\item{slr}{The slr data object to plot data from}

\item{year.above}{The year to start the boxplot}

\item{output}{Folder to put plots}

\item{save.pdf}{Whether to save or not plot to pdf file}
}
\description{
Title
}
