\name{bslr}
\alias{bslr}
\title{Balanced semi-Latin rectangles of cell size two}
\description{This function generates a balanced semi-Latin rectangles of cell size two}
\usage{bslr(v)}
\arguments{
	\item{v}{number of treatments}	
	}
\value{
	\item{design}{(Balanced semi-Latin rectangle of cell size two for v treatments}
	\item{Avg.Effi}{Average efficiency of the design}
  }
\examples{
bslr(5)
bslr(4)
}
\author{Kaushal Kumar Yadav <kaushalyadav0796@gmail.com>}
\keyword{balanced}
\keyword{semi-Latin rectangle}