% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{frequency_per_country}
\alias{frequency_per_country}
\title{Title}
\usage{
frequency_per_country(
  slr,
  year.above,
  output,
  save.pdf = T,
  plot.name = "SumCountries.pdf",
  width = 12.5
)
}
\arguments{
\item{slr}{The SLR data object to plot data from}

\item{year.above}{Filter by year the data on the SLR (default is above 1900)}

\item{output}{Folder where to write the plots into (default is the current directory - '.')}

\item{save.pdf}{Whether to save or not the plot to pdf file (default is TRUE)}

\item{plot.name}{Plot name}

\item{width}{Plot Width}
}
\description{
Title
}
