% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{frequency_per_institution}
\alias{frequency_per_institution}
\title{Title}
\usage{
frequency_per_institution(
  slr,
  year.above,
  output,
  save.pdf = T,
  frequency.above = 1,
  plot.name = "SumInstitutions.pdf",
  width = 12.5
)
}
\arguments{
\item{slr}{The SLR data object to plot data from}

\item{year.above}{Filter by year the data on the SLR (default is above 1900)}

\item{output}{Folder where to write the plots into (default is the current directory - '.')}

\item{save.pdf}{Whether to save or not the plot to pdf file (default is TRUE)}

\item{frequency.above}{Plot only Institutions with more than (n) studies}

\item{plot.name}{Plot name}

\item{width}{Plot Width}
}
\description{
Title
}
