% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{generate_slr_dimensions_plot}
\alias{generate_slr_dimensions_plot}
\title{Title}
\usage{
generate_slr_dimensions_plot(
  slr,
  year.above = 1900,
  output = ".",
  save.pdf = TRUE,
  dimensions
)
}
\arguments{
\item{slr}{The SLR data object to plot data from}

\item{year.above}{Filter by year the data on the SLR (default is above 1900)}

\item{output}{Folder where to write the plots into (default is the current directory - '.')}

\item{save.pdf}{Whether to save or not the plot to pdf file (default is TRUE)}

\item{dimensions}{List of dimensions to generate latex and plots}
}
\description{
Title
}
