% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_util.R
\name{getCoefsBM}
\alias{getCoefsBM}
\title{Get the coefficients of a balance regression model in terms of the linear log contrast model}
\usage{
getCoefsBM(coefs, sbp)
}
\arguments{
\item{coefs}{coefficients of balance regression model (may include intercept)}

\item{sbp}{fully-specified SBP vector/matrix}
}
\value{
the coefficients of a balance regression model in terms of the linear log contrast model
}
\description{
Get the coefficients of a balance regression model in terms of the linear log contrast model
}
