% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_util.R
\name{getSlrSBP}
\alias{getSlrSBP}
\title{Get a fully-specified SBP vector for an slr model}
\usage{
getSlrSBP(slr.sbp, x)
}
\arguments{
\item{slr.sbp}{sbp returned by slr object}

\item{x}{data matrix}
}
\value{
a fully-specified SBP vector
}
\description{
Get a fully-specified SBP vector for an slr model
}
