% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_util.R
\name{load_data}
\alias{load_data}
\title{Load the HIV data set}
\usage{
load_data()
}
\description{
Load the HIV data set from the selbal R package.
}
\details{
This function is a wrapper to import the \code{HIV} data set from the selbal R package.
Please make sure the package is installed by visiting \url{https://github.com/malucalle/selbal}.

\code{HIV} is a data frame with 155 rows (samples) and 62 columns (variables).
The first 60 variables measure the counts of bacterial species at the genus taxonomy rank.
The last column \code{HIV_Status} is a factor indicating the HIV infection status:
Pos/Neg if an individual is HIV1 positive/negative. There is also a column \code{MSM}
which is an HIV risk factor, Men who has sex with men (MSM) or not (nonMSM).
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/27077120/}
}
