% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_tests.R
\name{multiple_tests}
\alias{multiple_tests}
\title{Computes the tests between multiple combinations}
\usage{
multiple_tests(
  formula,
  data,
  FUN = "t.test",
  ...,
  response = "response",
  select = NULL,
  var_name = FALSE
)
}
\arguments{
\item{formula}{an object of class formula of the form y~factors or
y~factors|groups where y refers
to the response variable(s), factors determine the different test combination
and groups define the groups on which to carry the tests.}

\item{data}{a dataframe containing the variables defined in the formula}

\item{FUN}{the function f(x,y,...) used to carry out the test. eg t.test,
cor.test. default is t.test}

\item{...}{additional arguments to the function FUN above}

\item{response}{the resulting name that defines the different responses in
the formula}

\item{select}{the quantity needed from the test carried out eg. p.value etc}

\item{var_name}{logical indicating whether the factor name should be
included when generating the combinations}
}
\value{
a tidy dataframe containing all the quantities from the mutliple
tests
}
\description{
Computes the tests between multiple combinations
}
\examples{
multiple_tests(Sepal.Length~Species, iris)
# compare the first row to
t.test(Sepal.Length~Species, iris, subset = Species \%in\% c('setosa', 'versicolor'))
multiple_tests(cbind(Sepal.Length, Sepal.Width)~Species, iris)
multiple_tests(Sepal.Length + Sepal.Width~Species, iris)
multiple_tests(.~Species, iris)
multiple_tests(hp~am+vs, mtcars)
multiple_tests(qsec + hp~am+vs, mtcars, var_name = TRUE)
multiple_tests(hp~am|vs, mtcars, var_name = TRUE)
multiple_tests(.~Species, iris,cor.test)

}
