% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_simplelm_cpp.R
\name{my_simplelm_cpp}
\alias{my_simplelm_cpp}
\title{Manual Simple Linear Regression (C++ version)}
\usage{
my_simplelm_cpp(x, y, plot = TRUE)
}
\arguments{
\item{x}{Numeric vector (predictor)}

\item{y}{Numeric vector (response)}

\item{plot}{Logical. If TRUE, shows scatter plot with regression line.}
}
\value{
A list with Intercept, Slope, Fitted values, Residuals, and R-squared.
}
\description{
Uses Rcpp C++ code for fast linear regression.
}
\examples{
x <- 1:10
y <- 3 + 2*x + rnorm(10)
my_simplelm_cpp(x, y, plot = TRUE)
}
