% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dimension_frequency}
\alias{plot_dimension_frequency}
\title{Title}
\usage{
plot_dimension_frequency(
  dimension,
  slr,
  year.above = 1900,
  output = ".",
  save.pdf = TRUE,
  plot.name = "new_plot.pdf"
)
}
\arguments{
\item{dimension}{The dimension name to report against}

\item{slr}{The SLR data object to plot data from}

\item{year.above}{Filter by year the data on the SLR (default is above 1900)}

\item{output}{Folder where to write the plots into (default is the current directory - '.')}

\item{save.pdf}{Whether to save or not the plot to pdf file (default is TRUE)}

\item{plot.name}{Name of the file to plot data into}
}
\description{
Title
}
