% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{predict.slr}
\alias{predict.slr}
\title{Predict Method for slr Fits}
\usage{
\method{predict}{slr}(object, newdata = NULL, family = c("cox", "gaussian", "binomial"))
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{glm}}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{family}{Type of the response variable: could be \code{gaussian}, \code{binomial} or \code{cox}. This helps determine the prediction model.}
}
\value{
Predictions evaluated at \code{newdata}.
}
\description{
Obtain predictions from a fitted generalized linear model object
}
\details{
If \code{newdata} is omitted the predictions are based on the data used for the fit.
}
\examples{

HIV <- load_data() # Load HIV data
X <- HIV[,1:60]
y <- ifelse(HIV[,62] == "Pos", 1, 0)
X.adjusted <- sweep(X+1,rowSums(X+1),MARGIN = 1, FUN='/') # zero handling

# Run slr ----
fit <- slr(X.adjusted, y, method ='PC',
           family = 'binomial', threshold = 0.9)

predict.slr(fit,family = 'binomial')
}
