% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{read}
\alias{read}
\title{Title}
\usage{
read(filename, output = ".")
}
\arguments{
\item{filename}{the filename for the SLR data}

\item{output}{Folder where to write the Excel file with the articles (default is the current directory - '.')}
}
\description{
Title
}
