% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{score_per_publication_barchart}
\alias{score_per_publication_barchart}
\title{Title}
\usage{
score_per_publication_barchart(
  slr,
  year.above = 1900,
  output = ".",
  save.pdf = TRUE
)
}
\arguments{
\item{slr}{The SLR data object to plot data from}

\item{year.above}{Filter by year the data on the SLR (default is above 1900)}

\item{output}{Folder where to write the plots into (default is the current directory - '.')}

\item{save.pdf}{Whether to save or not the plot to pdf file (default is TRUE)}
}
\description{
Title
}
