% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr.fromContrast}
\alias{slr.fromContrast}
\title{Supervised Log-Ratios Microbial Signature}
\usage{
slr.fromContrast(x, contrast)
}
\arguments{
\item{x}{Matrix of relative abundances of compositional covariates.}

\item{contrast}{Contrast vector that takes values from -1, 0, 1, indicating the log-ratio discovered by slr.}
}
\value{
The microbial signature, or balance, corresponding to x, under the supervised log-ratios balance regression model.
}
\description{
Supervised Log-Ratios Microbial Signature
}
