% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{spectral.clustering}
\alias{spectral.clustering}
\title{Spectral Clustering}
\usage{
spectral.clustering(W, n_eig = 2, zeta = 0)
}
\arguments{
\item{W}{Similarity matrix.}

\item{n_eig}{Number of clusters.}

\item{zeta}{Small number used to perturb the network by adding some links with low edge weights in the calculation of the Normalized Laplacian.}
}
\value{
cluster assignments
}
\description{
Spectral Clustering
}
