% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_util.R
\docType{data}
\name{UC}
\alias{UC}
\title{Microbiome composition related to ulcerative colitis}
\format{
This data set is organized as a \code{data.frame} with the first 447 columns being genera counts, the next
column indicating the sample's status (UC vs Control), and the last column indicating the cohort.
}
\usage{
data(UC)
}
\description{
A processed microbiome data set consisting of the counts of 447 different genera in a group of 132 subjects.
There are two independent cohorts, which is given by the column \code{cohort}.
There are 34 controls and 53 UC subjects in the discovery cohort (\code{cohort=='PRISM'}),
and 22 controls and 23 UC subjects in the validation cohort (\code{cohort=='Validation'}).
For details on data preprocessing, please see documentation on GitHub \url{https://github.com/drjingma/LogRatioReg/tree/master}.
}
\references{
\url{https://doi.org/10.1038/s41564-018-0306-4}
}
\author{
Jing Ma (\email{jingma@fredhutch.org})
}
\keyword{data}
