% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{write_articles}
\alias{write_articles}
\title{Title}
\usage{
write_articles(slr, output = ".")
}
\arguments{
\item{slr}{The SLR data object to plot data from}

\item{output}{Folder where to write the latex files into (default is the current directory - '.')}
}
\description{
Title
}
