% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{write_comments}
\alias{write_comments}
\title{Write articles comments into a latex file, including the studi identifier and citation reference}
\usage{
write_comments(slr, rownames = TRUE, output = ".")
}
\arguments{
\item{slr}{The SLR data object to plot data from}

\item{rownames}{Flag to include rownames}

\item{output}{Folder where to write the latex files into (default is the current directory - '.')}
}
\description{
Write articles comments into a latex file, including the studi identifier and citation reference
}
