% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{write_graphics}
\alias{write_graphics}
\title{Title}
\usage{
write_graphics(output = ".")
}
\arguments{
\item{output}{Folder where to write the latex files into (default is the current directory - '.')}
}
\description{
Title
}
