% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{write_institutions}
\alias{write_institutions}
\title{Title}
\usage{
write_institutions(table, rownames = FALSE, output = ".")
}
\arguments{
\item{table}{The data table to render in latex}

\item{rownames}{Include or not the row names}

\item{output}{Folder where to write the latex files into (default is the current directory - '.')}
}
\description{
Title
}
