% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{writeComments}
\alias{writeComments}
\title{Title}
\usage{
writeComments(SLR_Final, rownames = TRUE, output = ".")
}
\arguments{
\item{SLR_Final}{Dataframe}

\item{rownames}{Flag to include rownames}

\item{output}{Folder to put plots}
}
\description{
Title
}
