% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{writeSummaryTable}
\alias{writeSummaryTable}
\title{Title}
\usage{
writeSummaryTable(table, rownames = TRUE, output = ".")
}
\arguments{
\item{table}{The data table to render in latex}

\item{rownames}{Include or not the row names}

\item{output}{Folder to output the latex file}
}
\description{
Title
}
