\name{bslr.even}
\alias{bslr.even}
\title{Balanced semi-Latin rectangles of cell size two for even number of treatments}
\description{This function generates a balanced semi-Latin rectangles of cell size two for even number of treatments}
\usage{bslr.even(v)}
\arguments{
	\item{v}{number of treatments and is even}	
	}
\value{
	\item{design}{(Balanced semi-Latin rectangle of cell size two for v treatments}
	\item{Avg.Effi}{Average efficiency of the design}
  }
\examples{
bslr.even(4)
}
\author{Sukanta Dash <sukanta.dash@icar.gov.in>}
\keyword{balanced}
\keyword{semi-Latin rectangle}