% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{cv.slr}
\alias{cv.slr}
\title{Cross Validation for Supervised Log Ratio}
\usage{
cv.slr(
  x,
  y,
  covar = NULL,
  family = c("cox", "gaussian", "binomial"),
  threshold = NULL,
  method = c("spectral clustering", "hierarchical clustering", "constrained PC"),
  zeta = 0,
  type.measure = c("default", "mse", "deviance", "class", "auc", "mae", "C", "accuracy"),
  nfolds = 5,
  foldid = NULL,
  weights = NULL,
  trace.it = FALSE,
  plot = FALSE
)
}
\arguments{
\item{x}{A sample by variable matrix of strictly positive relative abundances (\code{n} by \code{p}). This matrix should not contain any zeros.}

\item{y}{The response vector of length \code{n}.}

\item{covar}{A numeric matrix, containing additional covariates that you want to adjust for. If \code{NULL}, only feature abundances in \code{x} are used to predict the response variable.}

\item{family}{Type of the response variable: could be \code{gaussian}, \code{binomial} or \code{cox}.}

\item{threshold}{Nonnegative values between 0 and 1 used to screen variables. If \code{NULL}, the function will compute the threshold values internally.}

\item{method}{Method used to compute the principal balance. Could be \code{spectral clustering}, \code{hierarchical clustering} or \code{constrained PC}. See Details.}

\item{zeta}{A small positive value used to perturb the Aitchison similarity matrix when spectral clustering is used. Default is 0. See Details.}

\item{type.measure}{Loss used for cross-validation. If \code{family='gaussin'}, then \code{type.measure="mse"}. For two-class logistic regression, \code{type.measure="auc"}. For cox regression, \code{type.measure="C"}.}

\item{nfolds}{Number of folds. Default is 5.}

\item{foldid}{An optional vector of values between 1 and \code{nfold} identifying which fold each observation is in. If supplied, \code{nfold} can be missing.}

\item{weights}{Observation weights. Default is 1 per observation.}

\item{trace.it}{If \code{trace.it=TRUE}, then progress bars are displayed.}

\item{plot}{If \code{TRUE}, then a visualization of the cross-validation results is displayed.}
}
\value{
An object of class \code{"cv.slr"} is returned, which is a list with the ingredients of cross-validation fit.
\item{threshold}{A vector of threshold values used. The range of threshold values depends on the marginal association between each variable and the response.}
\item{cvm}{The mean cross-validated error - a vector of length \code{length(threshold)}.}
\item{cvsd}{The estimate of standard error of \code{cvm}.}
\item{foldid}{The fold assignments used.}
\item{threshold.min}{Value of \code{threshold} that gives minimum \code{cvm}.}
\item{threshold.1se}{Largest value of \code{threshoold} such that the cross-validation error is within 1 standard error of the minimum. This choice usually leads to a sparser set of selected variables.}
\item{index}{A one column matrix with the indices of \code{threshold.min} and \code{threshold.1se} in the sequence of all threshold values.}
}
\description{
Performs k-fold cross-validation for \code{slr} and returns an optimal value for \code{threshold}.
}
\details{
The function runs \code{slr} \code{nfolds} times to compute the fit with each of the folds omitted. The error is accumulated, and
the average error and standard deviation over the folds is computed. Note that the results of \code{cv.slr} are random, since the folds are selected at random.
Users can reduce this randomness by running \code{cv.slr} many times and averaging the error curves.

Please see \code{slr} for details regarding the choice of method and choice of \code{zeta}.
}
\examples{
library(compositions)
HIV <- load_data() # Load HIV data
X <- HIV[,1:60]
y <- ifelse(HIV[,62] == "Pos", 1, 0)
X.adjusted <- sweep(X+1,rowSums(X+1),MARGIN = 1, FUN='/')# zero handling

cv.out <- cv.slr(X.adjusted, y, method ='PC',
                 family = 'binomial',type.measure = 'auc',
                 trace.it = TRUE)
}
\seealso{
\code{slr}
}
\author{
Jing Ma and Kristyn Pantoja.

Maintainer: Jing Ma (\url{jingma@fredhutch.org})
}
