% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{cv.slr}
\alias{cv.slr}
\title{Cross Validation for Supervised Log-Ratios}
\usage{
cv.slr(
  x,
  y,
  screen.method = c("correlation", "wald"),
  cluster.method = c("spectral", "hierarchical"),
  response.type = c("survival", "continuous", "binary"),
  threshold = NULL,
  s0.perc = 0,
  zeta = 0,
  type.measure = c("default", "mse", "deviance", "class", "auc", "mae", "C",
    "accuracy"),
  nfolds = 10,
  foldid = NULL,
  weights = NULL,
  trace.it = FALSE
)
}
\arguments{
\item{x}{Matrix of relative abundances of compositional covariates (n sample proportions of p compositional components) for n observations with an observed response.}

\item{y}{Response vector (with length n).}

\item{screen.method}{Method of variable screening: could be correlation based ("correlation") or wald score based ("wald").}

\item{cluster.method}{Method of clustering: "spectral" or "hierarchical".}

\item{response.type}{Type of response variable: could be "survival", "continuous", or "binary". Currently only continuous and binary responses are allowed.}

\item{threshold}{Nonnegative constant between 0 and 1. If NULL, then no variable screening is performed.}

\item{s0.perc}{Factor for denominator of score statistic, between 0 and 1: the percentile of standard deviation values added to the denominator. Default is 0.}

\item{zeta}{Small number used to perturb the network by adding some links with low edge weights in the calculation of the Normalized Laplacian.}

\item{type.measure}{Loss used for cross-validation}

\item{nfolds}{Number of folds}

\item{foldid}{An optional vector of values between 1 and nfold identifying which fold each observation is in}

\item{weights}{Observation weights, default is 1}

\item{trace.it}{If trace.it=TRUE, then progress bars are displayed}
}
\value{
cv.slr object
}
\description{
Cross Validation for Supervised Log-Ratios
}
