% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr}
\alias{slr}
\title{Supervised Log-Ratios}
\usage{
slr(
  x,
  y,
  screen.method = c("correlation", "wald"),
  cluster.method = c("spectral", "hierarchical"),
  response.type = c("survival", "continuous", "binary"),
  threshold,
  s0.perc = 0,
  zeta = 0,
  positive.slope = FALSE
)
}
\arguments{
\item{x}{Matrix of relative abundances of compositional covariates (n sample proportions of p compositional components) for n observations with an observed response.}

\item{y}{Response vector (with length n).}

\item{screen.method}{Method of variable screening: could be correlation based ("correlation") or wald score based ("wald").}

\item{cluster.method}{Method of clustering: "spectral" or "hierarchical".}

\item{response.type}{Type of response variable: could be "survival", "continuous", or "binary". Currently only continuous and binary responses are allowed.}

\item{threshold}{Nonnegative constant between 0 and 1. If NULL, then no variable screening is performed.}

\item{s0.perc}{Factor for denominator of score statistic, between 0 and 1: the percentile of standard deviation values added to the denominator. Default is 0.}

\item{zeta}{Small number used to perturb the network by adding some links with low edge weights in the calculation of the Normalized Laplacian.}

\item{positive.slope}{Logical flag indicating whether to define the balance such that its corresponding parameter estimate is positive.}
}
\value{
slr object
}
\description{
Supervised Log-Ratios
}
