% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srl.R
\name{slr}
\alias{slr}
\title{Apply Simple Linear Regression}
\usage{
slr(x, y)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}
}
\value{
beta0,beta1,etc
}
\description{
Apply Simple Linear Regression
}
\examples{
library(slr)
slr(iris$Sepal.Length,iris$Sepal.Width)
}
